@@ -4,8 +4,8 @@ repository: https://github.com/OpenFlutter/flutter_oktoast
 version: 3.4.0
 
 environment:
-  sdk: '>=2.17.0 <4.0.0'
-  flutter: '>=3.7.0'
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: '>=3.10.0'
 
 dependencies:
   flutter:
